Convert civil time related types to time.Time
time.Time does not implement the scan
logic for civil time related types which the
google cloud boigquery SDK and hence this
gorm driver return.

For the go structs returned from the
gorm driver to be able to bind to
time.Time in downstream applications,
this PR changes the gorm driver to return
time.Time instead of civil types.